lab3: c: change RelType
Change RelType from size_t to int so release check works correctly.

+++ lab3/c/src/include/types.h
@@ -6,7 +6,7 @@
 // Types
 typedef int BusyType;
 typedef size_t KeyType;
-typedef size_t RelType;
+typedef int RelType;
 typedef size_t InfoType;
 typedef int IndexType;
